+++ app/Http/Controllers/App/TwoFactorAuthenticationController.php
@@ -27,7 +27,7 @@ class TwoFactorAuthenticationController extends Controller implements HasMiddlew
     {
         $request->ensureStateIsValid();
 
-        return inertia('/app/SettingTwoFactor', [
+        return inertia('app/SettingTwoFactor', [
             'twoFactorEnabled' => $request->user()->hasEnabledTwoFactorAuthentication(),
             'requiresConfirmation' => Features::optionEnabled(Features::twoFactorAuthentication(), 'confirm'),
         ]);
