+++ resources/js/layouts/app-layout/AppLogo.vue
@@ -7,6 +7,6 @@ import AppLogoIcon from '@/components/elements/AppLogoIcon.vue';
         <AppLogoIcon class="size-5 fill-current text-white dark:text-black" />
     </div>
     <div class="ml-1 grid flex-1 text-left text-sm">
-        <span class="mb-0.5 truncate leading-tight font-semibold">Laravel Starter Kit</span>
+        <span class="mb-0.5 truncate leading-tight font-semibold">Fintrack</span>
     </div>
 </template>
