+++ resources/js/layouts/app-layout/UserInfo.vue
@@ -1,8 +1,10 @@
 <script setup lang="ts">
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import { useInitials } from '@/composables/useInitials';
 import type { TUser } from '@/types/models';
+
+import { getInitials } from '@/lib/utils';
 import { computed } from 'vue';
+
+import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 
 interface Props {
     user: TUser;
@@ -13,9 +15,6 @@ const props = withDefaults(defineProps<Props>(), {
     showEmail: false,
 });
 
-const { getInitials } = useInitials();
-
-// Compute whether we should show the avatar image
 const showAvatar = computed(() => props.user.avatar && props.user.avatar !== '');
 </script>
