+++ resources/js/layouts/app-layout/UserMenuContent.vue
@@ -1,9 +1,11 @@
 <script setup lang="ts">
-import { DropdownMenuGroup, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator } from '@/components/ui/dropdown-menu';
-import { logout } from '@/routes';
-import { edit } from '@/routes/profile';
 import type { TUser } from '@/types/models';
-import { Link, router } from '@inertiajs/vue3';
+
+import { logout } from '@/routes';
+import { router } from '@inertiajs/vue3';
+
+import { DropdownMenuGroup, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator } from '@/components/ui/dropdown-menu';
+import { Link } from '@inertiajs/vue3';
 import { LogOut, Settings } from 'lucide-vue-next';
 
 import UserInfo from './UserInfo.vue';
@@ -28,7 +30,7 @@ defineProps<Props>();
     <DropdownMenuSeparator />
     <DropdownMenuGroup>
         <DropdownMenuItem :as-child="true">
-            <Link class="block w-full" :href="edit()" prefetch as="button">
+            <Link class="block w-full" :href="route('profile.edit')" prefetch as="button">
                 <Settings class="mr-2 h-4 w-4" />
                 Settings
             </Link>
